refactor: make measurex depend on measurexlite
